Refactor based on author's preferences
